// src/domain/content/enums.ts
export enum EContentType {
  MOVIE = 'movie',
  SERIES = 'series',
  DOCUMENTARY = 'documentary',
  ANIME = 'anime',
}

export enum EContentTier {
  FREE = 'free',
  PREMIUM = 'premium',
  PREMIUM_PLUS = 'premium_plus',
}

export enum EContentStatus {
  AVAILABLE = 'available',
  COMING_SOON = 'coming_soon',
  UNAVAILABLE = 'unavailable',
}